Refactor and expand wave charging and damping methods
Introduces new charging kernels (charge_falloff, charge_1lambda, charge_oscillator_falloff, charge_oscillator_wall) and envelope/damping control functions for more realistic and flexible wave initialization and energy management. Replaces and renames oscillate_spherical_standing with charge_oscillator_sphere, and comments out legacy wave center interaction code. Also comments out the absorbing boundary layer logic for future revision.
