+++ openwave/spacetime/E_wave_engine.py
@@ -147,13 +147,191 @@ def charge_gaussian(
         wave_field.psiT_old_am[i, j, k] = wave_field.psiT_am[i, j, k]
 
 
+@ti.kernel
+def charge_falloff(
+    wave_field: ti.template(),  # type: ignore
+    dt_rs: ti.f32,  # type: ignore
+):
+    """
+    Initialize a spherical outgoing wave with 1/r amplitude falloff.
+
+    Similar to initiate_charge() but includes realistic amplitude decay with
+    distance (λ/r falloff). Creates a radial sinusoidal displacement pattern
+    where amplitude decreases inversely with distance from the source.
+
+    Args:
+        wave_field: WaveField instance containing displacement arrays and grid info
+        dt_rs: Time step size (rs)
+    """
+
+    # Find center position (in grid indices)
+    center_x = wave_field.nx // 2
+    center_y = wave_field.ny // 2
+    center_z = wave_field.nz // 2
+
+    # Compute angular frequency (ω = 2πf) for temporal phase variation
+    omega_rs = (
+        2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
+    )  # angular frequency (rad/rs)
+
+    # Compute angular wave number (k = 2π/λ) for spatial phase variation
+    wavelength_grid = base_wavelength * wave_field.scale_factor / wave_field.dx
+    k_grid = 2.0 * ti.math.pi / wavelength_grid  # radians per grid index
+
+    # Create radial sinusoidal displacement pattern (interior points only)
+    # Skip boundaries to enforce Dirichlet boundary conditions (ψ=0 at edges)
+    for i, j, k in ti.ndrange(
+        (1, wave_field.nx - 1), (1, wave_field.ny - 1), (1, wave_field.nz - 1)
+    ):
+        # Distance from center (in grid indices)
+        r_grid = ti.sqrt((i - center_x) ** 2 + (j - center_y) ** 2 + (k - center_z) ** 2)
+
+        # Amplitude decreases with distance (1/r falloff)
+        r_safe_am = ti.max(r_grid, wavelength_grid)  # clamp to minimum 1λ to avoid singularity
+        amplitude_falloff = wavelength_grid / r_safe_am  # λ/r falloff factor
+        amplitude_am_at_r = base_amplitude_am * wave_field.scale_factor * amplitude_falloff
+
+        # Simple sinusoidal radial pattern
+        # Outward displacement from center source: A(r)·cos(ωt - kr)
+        # Creates rings of positive/negative displacement
+        # Signed value: positive = expansion, negative = compression
+        disp = amplitude_am_at_r * ti.cos(omega_rs * 0 - k_grid * r_grid)  # t0 initial condition
+        disp_old = amplitude_am_at_r * ti.cos(omega_rs * -dt_rs - k_grid * r_grid)
+
+        # Apply both displacements (in attometers)
+        wave_field.psiL_am[i, j, k] = disp  # at t=0
+        wave_field.psiL_old_am[i, j, k] = disp_old  # at t=-dt
+
+
+@ti.kernel
+def charge_1lambda(
+    wave_field: ti.template(),  # type: ignore
+    dt_rs: ti.f32,  # type: ignore
+):
+    """
+    Initialize a spherical outgoing wave within 1 wavelength.
+    Similar to initiate_charge() but limits the wave to within 1 wavelength
+
+    Args:
+        wave_field: WaveField instance containing displacement arrays and grid info
+        dt_rs: Time step size (rs)
+    """
+
+    # Find center position (in grid indices)
+    center_x = wave_field.nx // 2
+    center_y = wave_field.ny // 2
+    center_z = wave_field.nz // 2
+
+    # Compute angular frequency (ω = 2πf) for temporal phase variation
+    omega_rs = (
+        2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
+    )  # angular frequency (rad/rs)
+
+    # Compute angular wave number (k = 2π/λ) for spatial phase variation
+    wavelength_grid = base_wavelength * wave_field.scale_factor / wave_field.dx
+    k_grid = 2.0 * ti.math.pi / wavelength_grid  # radians per grid index
+
+    # Create radial sinusoidal displacement pattern (interior points only)
+    # Skip boundaries to enforce Dirichlet boundary conditions (ψ=0 at edges)
+    for i, j, k in ti.ndrange(
+        (1, wave_field.nx - 1), (1, wave_field.ny - 1), (1, wave_field.nz - 1)
+    ):
+        # Distance from center (in grid indices)
+        r_grid = ti.sqrt((i - center_x) ** 2 + (j - center_y) ** 2 + (k - center_z) ** 2)
+
+        # Amplitude = base if r < λ, 0 otherwise, oscillates radially
+        amplitude_am_at_r = (
+            base_amplitude_am * wave_field.scale_factor if r_grid < wavelength_grid else 0.0
+        )
+
+        # Simple sinusoidal radial pattern
+        # Outward displacement from center source: A(r)·cos(ωt - kr)
+        # Creates rings of positive/negative displacement
+        # Signed value: positive = expansion, negative = compression
+        disp = amplitude_am_at_r * ti.cos(omega_rs * 0 - k_grid * r_grid)  # t0 initial condition
+        disp_old = amplitude_am_at_r * ti.cos(omega_rs * -dt_rs - k_grid * r_grid)
+
+        # Apply both displacements (in attometers)
+        wave_field.psiL_am[i, j, k] = disp  # at t=0
+        wave_field.psiL_old_am[i, j, k] = disp_old  # at t=-dt
+
+
+# ================================================================
+# CHARGING CONTROL (smooth envelope for stability)
+# ================================================================
+
+
+def compute_charge_envelope(charge_level: float) -> float:
+    """
+    Compute smooth charging envelope based on current charge level.
+
+    HYBRID STRATEGY: Static pulse provides energy at initialization.
+    Dynamic chargers do the top-off from 70% to 100% with very soft landing.
+
+    Envelope curve:
+    - Full power (0-70%): Strong charging to reach target zone
+    - Soft taper (70-100%): Smooth S-curve for very gentle landing
+    - Off (100%+): No charging, let natural equilibrium hold
+
+    Args:
+        charge_level: Current energy as fraction of nominal (0.0 to 1.5+)
+
+    Returns:
+        float: Envelope value 0.0 to 1.0
+    """
+    TAPER_START = 0.70  # Start soft landing at 70%
+    TARGET = 1.00  # Target charge level
+
+    if charge_level >= TARGET:
+        return 0.0  # At/above target: no charging
+    elif charge_level >= TAPER_START:
+        # SOFT LANDING (70-100%): Smoothstep S-curve for very gentle approach
+        # Slower decay near target than cosine
+        t = (charge_level - TAPER_START) / (TARGET - TAPER_START)  # 0 at 70%, 1 at 100%
+        # Smoothstep: 3t² - 2t³ gives S-curve, invert for decay
+        smooth = t * t * (3.0 - 2.0 * t)  # 0→1 S-curve
+        return 1.0 - smooth  # 1.0 at 70%, 0.0 at 100%
+    else:
+        # FULL POWER (0-70%): Keep charging strong
+        return 1.0
+
+
+def compute_damping_factor(
+    charge_level: float, target: float = 1.0, tolerance: float = 0.10
+) -> float:
+    """
+    Compute proportional damping factor based on charge level.
+
+    HYBRID STRATEGY: Light baseline damping always to balance charger injection.
+    Stronger proportional damping above target to correct overshoots quickly.
+
+    Args:
+        charge_level: Current energy as fraction of nominal
+        target: Target charge level (default 1.0 = 100%)
+        tolerance: Overshoot tolerance before max damping (default 0.10 = 10%)
+
+    Returns:
+        float: Damping factor 0.98 to 0.9999
+    """
+    BASELINE = 0.99998  # Ultra light: ~0.002% per step, minimal drain
+
+    if charge_level <= target:
+        return BASELINE  # Light baseline damping
+    else:
+        # Above target: stronger proportional damping
+        overshoot = (charge_level - target) / tolerance
+        overshoot = min(overshoot, 5.0)
+        # Damping strength: BASELINE at target, 0.98 at target + 5*tolerance
+        return BASELINE - 0.004 * overshoot
+
+
 # ================================================================
 # DYNAMIC CHARGING methods (oscillator during simulation)
 # ================================================================
 
 
 @ti.kernel
-def oscillate_spherical_standing(
+def charge_oscillator_sphere(
     wave_field: ti.template(),  # type: ignore
     elapsed_t_rs: ti.f32,  # type: ignore
     radius: ti.f32,  # type: ignore
@@ -192,17 +370,152 @@ def oscillate_spherical_standing(
 
     # Apply oscillating displacement within source sphere
     # Harmonic motion: A·cos(ωt-kr), positive = expansion, negative = compression
-    nx, ny, nz = wave_field.nx, wave_field.ny, wave_field.nz
-    for i, j, k in ti.ndrange(nx, ny, nz):
+    for i, j, k in ti.ndrange(
+        (center_x - charge_radius_grid, center_x + charge_radius_grid + 1),
+        (center_y - charge_radius_grid, center_y + charge_radius_grid + 1),
+        (center_z - charge_radius_grid, center_z + charge_radius_grid + 1),
+    ):
         # Check if voxel is within spherical source region
         r_grid = ti.sqrt((i - center_x) ** 2 + (j - center_y) ** 2 + (k - center_z) ** 2)
-        wave_field.psiL_am[i, j, k] = (
-            base_amplitude_am
-            * boost
-            * wave_field.scale_factor
-            * ti.cos(omega_rs * elapsed_t_rs)
-            * ti.cos(k_grid * r_grid)
+        if r_grid <= charge_radius_grid:
+            wave_field.psiL_am[i, j, k] = (
+                base_amplitude_am
+                * boost
+                * wave_field.scale_factor
+                * ti.cos(omega_rs * elapsed_t_rs - k_grid * r_grid)
+            )
+
+
+@ti.kernel
+def charge_oscillator_falloff(
+    wave_field: ti.template(),  # type: ignore
+    elapsed_t_rs: ti.f32,  # type: ignore
+):
+    """
+    Apply harmonic oscillation with 1/r amplitude falloff.
+
+    Similar to charge_oscillator() but includes realistic amplitude decay with
+    distance (λ/r falloff). Creates a radial sinusoidal displacement pattern
+    where amplitude decreases inversely with distance from the source.
+
+    Creates a uniform displacement using:
+        ψ(t) = A(r)·cos(ωt-kr), where A(r) = A·(λ/r)
+
+    The oscillator acts as a coherent wave source, with all voxels inside
+    the sphere oscillating in phase.
+
+    Args:
+        wave_field: WaveField instance containing displacement arrays and grid info
+        elapsed_t_rs: Elapsed simulation time (rs)
+    """
+    # Compute angular frequency (ω = 2πf) for temporal phase variation
+    omega_rs = (
+        2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
+    )  # angular frequency (rad/rs)
+
+    # Compute angular wave number (k = 2π/λ) for spatial phase variation
+    wavelength_grid = base_wavelength * wave_field.scale_factor / wave_field.dx
+    k_grid = 2.0 * ti.math.pi / wavelength_grid  # radians per grid index
+
+    # Find center position (in grid indices)
+    center_x = wave_field.nx // 2
+    center_y = wave_field.ny // 2
+    center_z = wave_field.nz // 2
+
+    # Apply oscillating displacement
+    # Skip boundaries to enforce Dirichlet boundary conditions (ψ=0 at edges)
+    # Harmonic motion: A·cos(ωt-kr), positive = expansion, negative = compression
+    for i, j, k in ti.ndrange(
+        (1, wave_field.nx - 1), (1, wave_field.ny - 1), (1, wave_field.nz - 1)
+    ):
+        # Distance from center (in grid indices)
+        r_grid = ti.sqrt((i - center_x) ** 2 + (j - center_y) ** 2 + (k - center_z) ** 2)
+
+        # Amplitude decreases with distance (1/r falloff)
+        r_safe_am = ti.max(r_grid, wavelength_grid)  # clamp to minimum 1λ to avoid singularity
+        amplitude_falloff = wavelength_grid / r_safe_am  # λ/r falloff factor
+        amplitude_am_at_r = base_amplitude_am * wave_field.scale_factor * amplitude_falloff
+
+        wave_field.psiL_am[i, j, k] = amplitude_am_at_r * ti.cos(
+            omega_rs * elapsed_t_rs - k_grid * r_grid
         )
+
+
+@ti.kernel
+def charge_oscillator_wall(
+    wave_field: ti.template(),  # type: ignore
+    elapsed_t_rs: ti.f32,  # type: ignore
+    sources: ti.i32,  # type: ignore
+    boost: ti.f32,  # type: ignore
+):
+    """
+    Apply additive harmonic oscillation near boundary walls from source voxels.
+
+    ADDITIVE MODE: Adds oscillation to existing displacement rather than
+    overwriting. This prevents energy loss when propagated wave amplitude
+    exceeds charger amplitude, and allows natural wave superposition.
+
+    Formula: ψ_new = ψ_old + A·cos(ωt) * boost
+
+    Sources are placed 1 voxel interior from walls to avoid conflict with
+    Neumann BC ghost cell updates.
+
+    Args:
+        wave_field: WaveField instance containing displacement arrays and grid info
+        elapsed_t_rs: Elapsed simulation time (rs)
+        sources: Number of source points per edge
+        boost: Oscillation amplitude multiplier (envelope-controlled)
+    """
+    # Compute angular frequency (ω = 2πf) for temporal phase variation
+    omega_rs = (
+        2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
+    )  # angular frequency (rad/rs)
+
+    # Source grid calculation (interior, excluding boundary ghost cells)
+    # Range is [1, n-2] to stay 1 voxel away from boundaries
+    interior_nx = wave_field.nx - 2
+    interior_ny = wave_field.ny - 2
+    interior_nz = wave_field.nz - 2
+    sources_x = min(interior_nx, int(sources * interior_nx / wave_field.max_grid_size))
+    sources_y = min(interior_ny, int(sources * interior_ny / wave_field.max_grid_size))
+    sources_z = min(interior_nz, int(sources * interior_nz / wave_field.max_grid_size))
+    sources_x = max(sources_x, 1)
+    sources_y = max(sources_y, 1)
+    sources_z = max(sources_z, 1)
+    skip_x = interior_nx / sources_x
+    skip_y = interior_ny / sources_y
+    skip_z = interior_nz / sources_z
+    offset_x = skip_x / 2 + 1  # +1 to start at index 1, not 0
+    offset_y = skip_y / 2 + 1
+    offset_z = skip_z / 2 + 1
+
+    # Oscillation delta for this timestep (ADDITIVE, not absolute)
+    osc_delta = (
+        base_amplitude_am * boost * wave_field.scale_factor * ti.cos(omega_rs * elapsed_t_rs)
+    )
+
+    # ADD oscillating displacement on 6 near-boundary planes (1 voxel interior)
+    # Harmonic motion: ψ += A·cos(ωt), positive = expansion, negative = compression
+    # Z-faces: k=1 and k=nz-2
+    for i, j in ti.ndrange(sources_x, sources_y):
+        idx_x = int(i * skip_x + offset_x)
+        idx_y = int(j * skip_y + offset_y)
+        wave_field.psiL_am[idx_x, idx_y, 1] += osc_delta
+        wave_field.psiL_am[idx_x, idx_y, wave_field.nz - 2] += osc_delta
+
+    # Y-faces: j=1 and j=ny-2
+    for i, k in ti.ndrange(sources_x, sources_z):
+        idx_x = int(i * skip_x + offset_x)
+        idx_z = int(k * skip_z + offset_z)
+        wave_field.psiL_am[idx_x, 1, idx_z] += osc_delta
+        wave_field.psiL_am[idx_x, wave_field.ny - 2, idx_z] += osc_delta
+
+    # X-faces: i=1 and i=nx-2
+    for j, k in ti.ndrange(sources_y, sources_z):
+        idx_y = int(j * skip_y + offset_y)
+        idx_z = int(k * skip_z + offset_z)
+        wave_field.psiL_am[1, idx_y, idx_z] += osc_delta
+        wave_field.psiL_am[wave_field.nx - 2, idx_y, idx_z] += osc_delta
 
 
 # ================================================================
@@ -345,44 +658,44 @@ def propagate_wave(
             + (c_amrs * dt_rs) ** 2 * laplacianT_am
         )
 
-        # WAVE ABSORBING LAYER ============================================
-        # Reflection is inherent at wave equation boundaries:
-        # - Wave equation ∂²ψ/∂t² = c²∇²ψ is bidirectional by nature
-        # - Any impedance discontinuity (Z₁ ≠ Z₂) causes reflection: R = (Z₁-Z₂)/(Z₁+Z₂)
-        # - Sharp boundaries (Dirichlet, Neumann) are extreme impedance mismatches
-        # The solution = gradual impedance matching:
-        # - Quadratic damping profile creates smooth impedance transition
-        # - Wave "doesn't notice" it's being absorbed until it's too late
-        # - No sharp discontinuity = minimal reflection
-        # Energy conservation:
-        # - Damped psiT energy → psiL (not lost, just converted)
-        # - Physically meaningful: transverse component absorbed at "infinity" returns to longitudinal
-        # This is essentially a simplified PML (Perfectly Matched Layer) the standard technique
-        # in computational electromagnetics and acoustics for simulating infinite domains.
+        # # WAVE ABSORBING LAYER ============================================
+        # # Reflection is inherent at wave equation boundaries:
+        # # - Wave equation ∂²ψ/∂t² = c²∇²ψ is bidirectional by nature
+        # # - Any impedance discontinuity (Z₁ ≠ Z₂) causes reflection: R = (Z₁-Z₂)/(Z₁+Z₂)
+        # # - Sharp boundaries (Dirichlet, Neumann) are extreme impedance mismatches
+        # # The solution = gradual impedance matching:
+        # # - Quadratic damping profile creates smooth impedance transition
+        # # - Wave "doesn't notice" it's being absorbed until it's too late
+        # # - No sharp discontinuity = minimal reflection
+        # # Energy conservation:
+        # # - Damped psiT energy → psiL (not lost, just converted)
+        # # - Physically meaningful: transverse component absorbed at "infinity" returns to longitudinal
+        # # This is essentially a simplified PML (Perfectly Matched Layer) the standard technique
+        # # in computational electromagnetics and acoustics for simulating infinite domains.
 
-        # Calculate distance to nearest boundary for absorbing layer
-        dist_x = ti.min(i, nx - 1 - i)
-        dist_y = ti.min(j, ny - 1 - j)
-        dist_z = ti.min(k, nz - 1 - k)
-        dist_to_boundary = ti.min(dist_x, ti.min(dist_y, dist_z))
+        # # Calculate distance to nearest boundary for absorbing layer
+        # dist_x = ti.min(i, nx - 1 - i)
+        # dist_y = ti.min(j, ny - 1 - j)
+        # dist_z = ti.min(k, nz - 1 - k)
+        # dist_to_boundary = ti.min(dist_x, ti.min(dist_y, dist_z))
 
-        # Absorbing layer: gentle damping applied AFTER wave equation
-        # Small damping with quadratic profile for gradual impedance change
-        absorbing_width = nx // 10  # fraction of grid on each side
-        if dist_to_boundary < absorbing_width:
-            normalized_dist = (absorbing_width - dist_to_boundary) / absorbing_width
-            damping = 0.10 * normalized_dist**2  # Max 10% at boundary
+        # # Absorbing layer: gentle damping applied AFTER wave equation
+        # # Small damping with quadratic profile for gradual impedance change
+        # absorbing_width = nx // 10  # fraction of grid on each side
+        # if dist_to_boundary < absorbing_width:
+        #     normalized_dist = (absorbing_width - dist_to_boundary) / absorbing_width
+        #     damping = 0.10 * normalized_dist**2  # Max 10% at boundary
 
-            psiT_before = wave_field.psiT_new_am[i, j, k]
-            psiT_after = psiT_before * (1.0 - damping)
-            wave_field.psiT_new_am[i, j, k] = psiT_after
-            wave_field.psiL_new_am[i, j, k] = wave_field.psiL_new_am[i, j, k] * (1.0 - damping)
+        #     psiT_before = wave_field.psiT_new_am[i, j, k]
+        #     psiT_after = psiT_before * (1.0 - damping)
+        #     wave_field.psiT_new_am[i, j, k] = psiT_after
+        #     wave_field.psiL_new_am[i, j, k] = wave_field.psiL_new_am[i, j, k] * (1.0 - damping)
 
-            # Transfer damped psiT energy to psiL (branchless)
-            # energy_diff = ti.max(0.0, psiT_before**2 - psiT_after**2)
-            # psiL_before = wave_field.psiL_new_am[i, j, k]
-            # psiL_sign = ti.select(psiL_before >= 0.0, 1.0, -1.0)
-            # wave_field.psiL_new_am[i, j, k] = psiL_sign * ti.sqrt(psiL_before**2 + energy_diff)
+        #     # Transfer damped psiT energy to psiL (branchless)
+        #     # energy_diff = ti.max(0.0, psiT_before**2 - psiT_after**2)
+        #     # psiL_before = wave_field.psiL_new_am[i, j, k]
+        #     # psiL_sign = ti.select(psiL_before >= 0.0, 1.0, -1.0)
+        #     # wave_field.psiL_new_am[i, j, k] = psiL_sign * ti.sqrt(psiL_before**2 + energy_diff)
 
         # WAVE-TRACKERS ============================================
         # RMS AMPLITUDE tracking via EMA on ψ² (squared displacement)
@@ -433,694 +746,37 @@ def propagate_wave(
         wave_field.psiT_old_am[i, j, k] = wave_field.psiT_am[i, j, k]
         wave_field.psiT_am[i, j, k] = wave_field.psiT_new_am[i, j, k]
 
-    # TODO: Testing Wave Center Interaction with Energy Waves
-    # WCs modify Energy Wave character (amplitude/phase/lambda/mode) as they pass through
-    # Standing Waves should form around WCs as visual artifacts of interaction
-    # Energy Waves are Isotropic (omnidirectional) so reflection gets canceled out
-
-    # interact_wc1_pulse(wave_field, elapsed_t_rs)  # forces amp, but no standing wave formation
-    # interact_wc2_pulse(wave_field, elapsed_t_rs)  # forces amp, but no standing wave formation
-
-    # wc1_standing(wave_field, elapsed_t_rs, sim_speed)
-    # wc2_standing(wave_field, elapsed_t_rs, sim_speed)
-
-    # interact_wc_spinUP(wave_field, dt_rs)  # never worked correctly
-    # interact_wc_spinUP2(wave_field, dt_rs)  # never worked correctly
-    # interact_wc_spinDOWN(wave_field, dt_rs)  # never worked correctly
-
-    # interact_wc_swap(wave_field)
-    # interact_wc_lens(wave_field)  # amplify waves at WC, but no standing wave formation
-
-    # interact_wc_min(wave_field) # no interaction on isotropic waves
-    # interact_wc_drain(wave_field) # no interaction on isotropic waves
-    # interact_wc_newmann(wave_field) # no interaction on isotropic waves
-    # interact_wc_dirichlet(wave_field) # no interaction on isotropic waves
-    # interact_wc_signal(wave_field) # just invert oscillation, no wave interaction
-
 
 # ================================================================
-# WAVE CENTER INTERACTIONS
+# DYNAMIC DAMPING methods (energy sink during simulation)
 # ================================================================
 
 
-@ti.func
-def interact_wc1_pulse(wave_field: ti.template(), elapsed_t_rs):  # type: ignore
+@ti.kernel
+def damp_full(
+    wave_field: ti.template(),  # type: ignore
+    damping_factor: ti.f32,  # type: ignore
+):
     """
-    TEST: Wave center as PULSE - injects oscillation at WC sphere surface
+    Gradually absorb wave energy within the entire grid volume.
+
+    Applies exponential damping to displacement values, simulating energy
+    absorption. Each frame, displacement is multiplied by damping_factor,
+    causing gradual decay toward zero.
 
     Args:
         wave_field: WaveField instance containing displacement arrays and grid info
+        damping_factor: Damping multiplier per frame (e.g., 0.95 = 5% absorption per frame)
     """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-    wc_radius = 0  # radius in voxels
-    wc_radius_sq = wc_radius**2  # radius² = 8² voxels (≈ λ/4 for λ=30dx)
-    boost = 1  # amplitude boost factor
 
-    # Compute angular frequency (ω = 2πf) for temporal phase variation
-    omega_rs = (
-        2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
-    )  # angular frequency (rad/rs)
-
+    # Apply damping displacement within entire grid
+    # NOTE: Must be called AFTER propagate_ewave to damp the propagated values
     for i, j, k in ti.ndrange(
-        (wc1x - wc_radius - 1, wc1x + wc_radius + 2),
-        (wc1y - wc_radius - 1, wc1y + wc_radius + 2),
-        (wc1z - wc_radius - 1, wc1z + wc_radius + 2),
+        (0, wave_field.nx),
+        (0, wave_field.ny),
+        (0, wave_field.nz),
     ):
-        dist_sq = (i - wc1x) ** 2 + (j - wc1y) ** 2 + (k - wc1z) ** 2
-        # Only process voxels on inner surface of sphere (r = radius-1)
-        if dist_sq <= wc_radius_sq:
-            wave_field.psiL_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs)
-            )
-            wave_field.psiT_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs)
-            )
-
-
-@ti.func
-def interact_wc2_pulse(wave_field: ti.template(), elapsed_t_rs):  # type: ignore
-    """
-    TEST: Wave center as PULSE - injects oscillation at WC sphere surface
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-    """
-    wc2x, wc2y, wc2z = wave_field.nx * 9 // 12, wave_field.ny * 9 // 12, wave_field.nz // 2
-    wc_radius = 0  # radius in voxels
-    wc_radius_sq = wc_radius**2  # radius² = 8² voxels (≈ λ/4 for λ=30dx)
-    boost = 1  # amplitude boost factor
-
-    # Compute angular frequency (ω = 2πf) for temporal phase variation
-    omega_rs = (
-        2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
-    )  # angular frequency (rad/rs)
-
-    for i, j, k in ti.ndrange(
-        (wc2x - wc_radius - 1, wc2x + wc_radius + 2),
-        (wc2y - wc_radius - 1, wc2y + wc_radius + 2),
-        (wc2z - wc_radius - 1, wc2z + wc_radius + 2),
-    ):
-        dist_sq = (i - wc2x) ** 2 + (j - wc2y) ** 2 + (k - wc2z) ** 2
-        # Only process voxels on inner surface of sphere (r = radius-1)
-        if dist_sq <= wc_radius_sq:
-            wave_field.psiL_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs)
-            )
-            wave_field.psiT_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs + ti.math.pi)  # 180° phase shift
-            )
-
-
-@ti.func
-def wc1_standing(wave_field: ti.template(), elapsed_t_rs, sim_speed: ti.f32):  # type: ignore
-    """
-    TEST: Wave center as STANDING WAVE - injects oscillation at WC sphere surface
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-        elapsed_t_rs: Elapsed time in reduced seconds
-        sim_speed: Simulation speed factor
-    """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-    wc_radius = int(wave_field.ewave_res * 1)  # radius in voxels
-    wc_radius_sq = wc_radius**2  # radius² = 8² voxels (≈ λ/4 for λ=30dx)
-    boost = 1.0  # amplitude boost factor
-
-    # Compute angular frequency (ω = 2πf) for temporal phase variation
-    omega_rs = (
-        (1 * sim_speed / boost) * 2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
-    )  # angular frequency (rad/rs)
-
-    # Compute angular wave number (k = 2π/λ) for spatial phase variation
-    wavelength_grid = base_wavelength * wave_field.scale_factor / wave_field.dx  # voxels / λ
-    k_grid = 2.0 * ti.math.pi / wavelength_grid  # radians per grid index
-
-    for i, j, k in ti.ndrange(
-        (wc1x - wc_radius - 1, wc1x + wc_radius + 2),
-        (wc1y - wc_radius - 1, wc1y + wc_radius + 2),
-        (wc1z - wc_radius - 1, wc1z + wc_radius + 2),
-    ):
-        dist_sq = (i - wc1x) ** 2 + (j - wc1y) ** 2 + (k - wc1z) ** 2
-        r_grid = ti.sqrt(dist_sq)  # Distance from wave-center (in grid indices)
-        # Only process voxels on inner surface of sphere (r = radius-1)
-        if dist_sq <= wc_radius_sq and dist_sq > 0:
-            wave_field.psiL_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs)
-                * ti.sin(k_grid * r_grid)
-                / r_grid
-            )
-            wave_field.psiT_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs)
-                * ti.sin(k_grid * r_grid)
-                / r_grid
-            )
-
-        elif dist_sq == 0:
-            wave_field.psiL_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs)
-                * 1
-            )  # Avoids singularity at the wave-center
-            wave_field.psiT_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs)
-                * 1
-            )  # Avoids singularity at the wave-center
-
-
-@ti.func
-def wc2_standing(wave_field: ti.template(), elapsed_t_rs, sim_speed: ti.f32):  # type: ignore
-    """
-    TEST: Wave center as STANDING WAVE - injects oscillation at WC sphere surface
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-    """
-    wc2x, wc2y, wc2z = wave_field.nx * 9 // 12, wave_field.ny * 9 // 12, wave_field.nz // 2
-    wc_radius = int(wave_field.ewave_res * 1)  # radius in voxels
-    wc_radius_sq = wc_radius**2  # radius² = 8² voxels (≈ λ/4 for λ=30dx)
-    boost = 1.0  # amplitude boost factor
-
-    # Compute angular frequency (ω = 2πf) for temporal phase variation
-    omega_rs = (
-        (1 * sim_speed / boost) * 2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
-    )  # angular frequency (rad/rs)
-
-    # Compute angular wave number (k = 2π/λ) for spatial phase variation
-    wavelength_grid = base_wavelength * wave_field.scale_factor / wave_field.dx  # voxels / λ
-    k_grid = 2.0 * ti.math.pi / wavelength_grid  # radians per grid index
-
-    for i, j, k in ti.ndrange(
-        (wc2x - wc_radius - 1, wc2x + wc_radius + 2),
-        (wc2y - wc_radius - 1, wc2y + wc_radius + 2),
-        (wc2z - wc_radius - 1, wc2z + wc_radius + 2),
-    ):
-        dist_sq = (i - wc2x) ** 2 + (j - wc2y) ** 2 + (k - wc2z) ** 2
-        r_grid = ti.sqrt(dist_sq)  # Distance from wave-center (in grid indices)
-        # Only process voxels on inner surface of sphere (r = radius-1)
-        if dist_sq <= wc_radius_sq and dist_sq > 0:
-            wave_field.psiL_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs)
-                * ti.sin(k_grid * r_grid)
-                / r_grid
-            )
-            wave_field.psiT_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs + ti.math.pi)
-                * ti.sin(k_grid * r_grid)
-                / r_grid
-            )
-        elif dist_sq == 0:
-            wave_field.psiL_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs)
-                * 1
-            )  # Avoids singularity at the wave-center
-            wave_field.psiT_am[i, j, k] = (
-                base_amplitude_am
-                * boost
-                * wave_field.scale_factor
-                * ti.cos(omega_rs * elapsed_t_rs + ti.math.pi)
-                * 1
-            )  # Avoids singularity at the wave-center
-
-
-@ti.func
-def interact_wc_spinUP(wave_field: ti.template(), dt_rs: ti.f32):  # type: ignore
-    """
-    Wave center spin-UP interaction: phase-shifts psiL by +90° and creates psiT.
-
-    WAVE CENTER SPIN MECHANISM:
-    1. Incoming longitudinal wave psiL contacts wave center
-    2. WC spin creates transverse component: psiT = α × psiL (fine structure ratio)
-    3. Outgoing psiL is PHASE-SHIFTED by +90° (counterclockwise/leading)
-
-    PHASE SHIFT VIA VELOCITY:
-    For sinusoidal wave psiL = A·cos(ωt):
-        - Velocity: ∂psiL/∂t = -A·ω·sin(ωt)
-        - Normalized: velocity/ω = -A·sin(ωt) = A·cos(ωt + 90°)  ← +90° shifted!
-
-    So (psiL - psiL_old)/(ω·dt) gives the +90° phase-shifted wave.
-    This creates a DISTURBANCE in the longitudinal wave from the spin interaction.
-
-    ENERGY CONSERVATION:
-        psiT = α × psiL          (transverse component created)
-        psiL_out² + psiT² = psiL_in²  (total energy conserved)
-        psiL_out = ±√(psiL² - psiT²)  with sign from phase-shifted wave
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-        dt_rs: Time step size (rs) for velocity calculation
-    """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-    alpha = constants.FINE_STRUCTURE  # L→T conversion ratio
-
-    # Angular frequency (scaled for simulation)
-    omega_rs = 2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
-
-    # Current and previous longitudinal displacement at WC
-    psiL = wave_field.psiL_am[wc1x, wc1y, wc1z]
-    psiL_old = wave_field.psiL_old_am[wc1x, wc1y, wc1z]
-
-    # ================================================================
-    # STEP 1: Compute phase-shifted psiL (+90° leading)
-    # ================================================================
-    # Velocity via finite difference
-    delta_psiL = psiL - psiL_old
-
-    # Phase-shifted psiL: -velocity/ω = -delta_psiL / (ω·dt)
-    # delta_psiL ≈ -ω·dt·sin(ωt) for cos input, so:
-    # psiL_shifted = -(-ω·dt·sin)/(ω·dt) = +sin(ωt)
-    # This gives +90° shift: cos → sin
-    psiL_shifted = -delta_psiL / (omega_rs * dt_rs)
-
-    # ================================================================
-    # STEP 2: Create transverse component (90° from psiL_out)
-    # ================================================================
-    # We've tried:
-    #   psiT = alpha * psiL → gives 180°
-    #   psiT = alpha * psiL_shifted → gives 0°
-    # For 90°, try negating one: psiT = -alpha * psiL
-    psiT = -alpha * psiL  # NEGATED to flip 180° → hopefully 90°
-
-    # Safety clamp: ensure psiT² < psiL² to prevent NaN from sqrt
-    max_psiT = 0.99 * ti.abs(psiL)
-    psiT = ti.math.clamp(psiT, -max_psiT, max_psiT)
-
-    # ================================================================
-    # STEP 3: Output psiL as actual phase-shifted wave
-    # ================================================================
-    # Energy available for psiL_out after psiT extraction
-    psiL_energy = psiL**2 - psiT**2
-
-    # Scale psiL_shifted to have correct energy while preserving its phase
-    psiL_shifted_sq = psiL_shifted**2
-
-    # Initialize psiL_out (required for Taichi scoping)
-    psiL_out = 0.0
-
-    # Avoid division by zero: if psiL_shifted is tiny, fall back to original
-    if psiL_shifted_sq > 1e-20:
-        scaling = ti.sqrt(psiL_energy / psiL_shifted_sq)
-        psiL_out = psiL_shifted * scaling
-    else:
-        # At zero crossing of shifted wave, preserve sign of original
-        phase_sign = 1.0 if psiL >= 0.0 else -1.0
-        psiL_out = phase_sign * ti.sqrt(psiL_energy)
-
-    wave_field.psiL_am[wc1x, wc1y, wc1z] = psiL_out
-    wave_field.psiT_am[wc1x, wc1y, wc1z] = psiT
-
-
-@ti.func
-def interact_wc_spinUP2(wave_field: ti.template(), dt_rs: ti.f32):  # type: ignore
-    """
-    Wave center spin-UP interaction: phase-shifts psiL by +90° and creates psiT.
-
-    WAVE CENTER SPIN MECHANISM:
-    1. Incoming longitudinal wave psiL contacts wave center
-    2. WC spin creates transverse component: psiT = α × psiL (fine structure ratio)
-    3. Outgoing psiL is PHASE-SHIFTED by +90° (counterclockwise/leading)
-
-    PHASE SHIFT VIA VELOCITY:
-    For sinusoidal wave psiL = A·cos(ωt):
-        - Velocity: ∂psiL/∂t = -A·ω·sin(ωt)
-        - Normalized: velocity/ω = -A·sin(ωt) = A·cos(ωt + 90°)  ← +90° shifted!
-
-    So (psiL - psiL_old)/(ω·dt) gives the +90° phase-shifted wave.
-    This creates a DISTURBANCE in the longitudinal wave from the spin interaction.
-
-    ENERGY CONSERVATION:
-        psiT = α × psiL          (transverse component created)
-        psiL_out² + psiT² = psiL_in²  (total energy conserved)
-        psiL_out = ±√(psiL² - psiT²)  with sign from phase-shifted wave
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-        dt_rs: Time step size (rs) for velocity calculation
-    """
-    wc2x, wc2y, wc2z = wave_field.nx * 9 // 12, wave_field.ny * 9 // 12, wave_field.nz // 2
-    alpha = constants.FINE_STRUCTURE  # L→T conversion ratio
-
-    # Angular frequency (scaled for simulation)
-    omega_rs = 2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
-
-    # Current and previous longitudinal displacement at WC
-    psiL = wave_field.psiL_am[wc2x, wc2y, wc2z]
-    psiL_old = wave_field.psiL_old_am[wc2x, wc2y, wc2z]
-
-    # ================================================================
-    # STEP 1: Compute phase-shifted psiL (+90° leading)
-    # ================================================================
-    # Velocity via finite difference
-    delta_psiL = psiL - psiL_old
-
-    # Phase-shifted psiL: -velocity/ω = -delta_psiL / (ω·dt)
-    # delta_psiL ≈ -ω·dt·sin(ωt) for cos input, so:
-    # psiL_shifted = -(-ω·dt·sin)/(ω·dt) = +sin(ωt)
-    # This gives +90° shift: cos → sin
-    psiL_shifted = -delta_psiL / (omega_rs * dt_rs)
-
-    # ================================================================
-    # STEP 2: Create transverse component (90° from psiL_out)
-    # ================================================================
-    # We've tried:
-    #   psiT = alpha * psiL → gives 180°
-    #   psiT = alpha * psiL_shifted → gives 0°
-    # For 90°, try negating one: psiT = -alpha * psiL
-    psiT = -alpha * psiL  # NEGATED to flip 180° → hopefully 90°
-
-    # Safety clamp: ensure psiT² < psiL² to prevent NaN from sqrt
-    max_psiT = 0.99 * ti.abs(psiL)
-    psiT = ti.math.clamp(psiT, -max_psiT, max_psiT)
-
-    # ================================================================
-    # STEP 3: Output psiL as actual phase-shifted wave
-    # ================================================================
-    # Energy available for psiL_out after psiT extraction
-    psiL_energy = psiL**2 - psiT**2
-
-    # Scale psiL_shifted to have correct energy while preserving its phase
-    psiL_shifted_sq = psiL_shifted**2
-
-    # Initialize psiL_out (required for Taichi scoping)
-    psiL_out = 0.0
-
-    # Avoid division by zero: if psiL_shifted is tiny, fall back to original
-    if psiL_shifted_sq > 1e-20:
-        scaling = ti.sqrt(psiL_energy / psiL_shifted_sq)
-        psiL_out = psiL_shifted * scaling
-    else:
-        # At zero crossing of shifted wave, preserve sign of original
-        phase_sign = 1.0 if psiL >= 0.0 else -1.0
-        psiL_out = phase_sign * ti.sqrt(psiL_energy)
-
-    wave_field.psiL_am[wc2x, wc2y, wc2z] = psiL_out
-    wave_field.psiT_am[wc2x, wc2y, wc2z] = psiT
-
-
-@ti.func
-def interact_wc_spinDOWN(wave_field: ti.template(), dt_rs: ti.f32):  # type: ignore
-    """
-    Wave center spin-DOWN interaction: phase-shifts psiL by -90° and creates psiT.
-
-    WAVE CENTER SPIN MECHANISM (opposite direction):
-    1. Incoming longitudinal wave psiL contacts wave center
-    2. WC spin creates transverse component: psiT = α × psiL (fine structure ratio)
-    3. Outgoing psiL is PHASE-SHIFTED by -90° (clockwise/lagging)
-
-    PHASE SHIFT VIA NEGATIVE VELOCITY:
-    For sinusoidal wave psiL = A·cos(ωt):
-        - Velocity: ∂psiL/∂t = -A·ω·sin(ωt)
-        - Negative normalized: -velocity/ω = A·sin(ωt) = A·cos(ωt - 90°)  ← -90° shifted!
-
-    So -(psiL - psiL_old)/(ω·dt) gives the -90° phase-shifted wave.
-    This creates a DISTURBANCE in the longitudinal wave (opposite to spinUP).
-
-    ENERGY CONSERVATION:
-        psiT = α × psiL          (transverse component created)
-        psiL_out² + psiT² = psiL_in²  (total energy conserved)
-        psiL_out = ±√(psiL² - psiT²)  with sign from phase-shifted wave
-
-    COMPARISON:
-        spinUP:   psiL phase +90° (leading),  counterclockwise
-        spinDOWN: psiL phase -90° (lagging),  clockwise
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-        dt_rs: Time step size (rs) for velocity calculation
-    """
-    wc2x, wc2y, wc2z = wave_field.nx * 9 // 12, wave_field.ny * 9 // 12, wave_field.nz // 2
-    alpha = constants.FINE_STRUCTURE  # L→T conversion ratio
-
-    # Angular frequency (scaled for simulation)
-    omega_rs = 2.0 * ti.math.pi * base_frequency_rHz / wave_field.scale_factor
-
-    # Current and previous longitudinal displacement at WC
-    psiL = wave_field.psiL_am[wc2x, wc2y, wc2z]
-    psiL_old = wave_field.psiL_old_am[wc2x, wc2y, wc2z]
-
-    # ================================================================
-    # STEP 1: Compute phase-shifted psiL (-90° lagging)
-    # ================================================================
-    # Velocity via finite difference
-    delta_psiL = psiL - psiL_old
-
-    # Phase-shifted psiL: -velocity/ω = -delta_psiL / (ω·dt)
-    # Transforms A·cos(ωt) → A·sin(ωt) = A·cos(ωt - 90°)
-    psiL_shifted = -delta_psiL / (omega_rs * dt_rs)
-
-    # ================================================================
-    # STEP 2: Create transverse component (90° from psiL_out)
-    # ================================================================
-    # Negated to achieve 90° phase relationship
-    psiT = -alpha * psiL  # NEGATED
-
-    # Safety clamp: ensure psiT² < psiL² to prevent NaN from sqrt
-    max_psiT = 0.99 * ti.abs(psiL)
-    psiT = ti.math.clamp(psiT, -max_psiT, max_psiT)
-
-    # ================================================================
-    # STEP 3: Output psiL as actual phase-shifted wave
-    # ================================================================
-    # Energy available for psiL_out after psiT extraction
-    psiL_energy = psiL**2 - psiT**2
-
-    # Scale psiL_shifted to have correct energy while preserving its phase
-    psiL_shifted_sq = psiL_shifted**2
-
-    # Initialize psiL_out (required for Taichi scoping)
-    psiL_out = 0.0
-
-    # Avoid division by zero: if psiL_shifted is tiny, fall back to original
-    if psiL_shifted_sq > 1e-20:
-        scaling = ti.sqrt(psiL_energy / psiL_shifted_sq)
-        psiL_out = psiL_shifted * scaling
-    else:
-        # At zero crossing of shifted wave, preserve sign of original
-        phase_sign = 1.0 if psiL >= 0.0 else -1.0
-        psiL_out = phase_sign * ti.sqrt(psiL_energy)
-
-    wave_field.psiL_am[wc2x, wc2y, wc2z] = psiL_out
-    wave_field.psiT_am[wc2x, wc2y, wc2z] = psiT
-
-
-@ti.func
-def interact_wc_swap(wave_field: ti.template()):  # type: ignore
-    """
-    TEST: Wave center swaps neighbor displacements.
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-    """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-
-    # Swap displacement at X axis
-    left = wave_field.psiL_am[wc1x + 1, wc1y, wc1z]
-    right = wave_field.psiL_am[wc1x - 1, wc1y, wc1z]
-    wave_field.psiL_am[wc1x + 1, wc1y, wc1z] = right
-    wave_field.psiL_am[wc1x - 1, wc1y, wc1z] = left
-
-    # Swap displacement at Y axis
-    front = wave_field.psiL_am[wc1x, wc1y + 1, wc1z]
-    back = wave_field.psiL_am[wc1x, wc1y - 1, wc1z]
-    wave_field.psiL_am[wc1x, wc1y + 1, wc1z] = back
-    wave_field.psiL_am[wc1x, wc1y - 1, wc1z] = front
-
-    # Swap displacement at Z axis
-    top = wave_field.psiL_am[wc1x, wc1y, wc1z + 1]
-    bottom = wave_field.psiL_am[wc1x, wc1y, wc1z - 1]
-    wave_field.psiL_am[wc1x, wc1y, wc1z + 1] = bottom
-    wave_field.psiL_am[wc1x, wc1y, wc1z - 1] = top
-
-
-@ti.func
-def interact_wc_lens(wave_field: ti.template()):  # type: ignore
-    """
-    TEST: Wave center as LENS - amplitude increase
-    Absorbs wave energy at the wave center (WC) sphere surface
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-    """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-    wc_radius = 1  # radius in voxels
-    wc_radius_sq = wc_radius**2  # radius² = 8² voxels (≈ λ/4 for λ=30dx)
-    l = 2  # lens amplification factor
-
-    for i, j, k in ti.ndrange(
-        (wc1x - wc_radius - 1, wc1x + wc_radius + 2),
-        (wc1y - wc_radius - 1, wc1y + wc_radius + 2),
-        (wc1z - wc_radius - 1, wc1z + wc_radius + 2),
-    ):
-        dist_sq = (i - wc1x) ** 2 + (j - wc1y) ** 2 + (k - wc1z) ** 2
-        # Only process voxels on inner surface of sphere (r = radius-1)
-        if dist_sq <= wc_radius_sq:
-            wave_field.psiL_am[i, j, k] = l * wave_field.psiL_am[i, j, k]
-
-
-@ti.func
-def interact_wc_min(wave_field: ti.template()):  # type: ignore
-    """
-    TEST: Wave center as amplification
-    Amplifies/focuses waves rather than blocking
-    Allow normal wave propagation but clamp the minimum amplitude at WC
-    WC is a point-like region where wave amplitude is amplified
-    This creates a phase singularity that surrounding waves conform to
-    If |ψ| < threshold, boost it to threshold (preserving sign)
-    This creates an amplitude floor, not ceiling — no runaway growth
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-    """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-    amplification = 2.0  # how much the WC amplifies local wave amplitude
-
-    # Reference amplitude from wave field (the base amplitude used for charging)
-    ref_amplitude = base_amplitude_am * wave_field.scale_factor
-    min_amplitude = ref_amplitude * amplification
-
-    # If amplitude is below minimum, boost it (preserve sign/phase)
-    current_val = wave_field.psiL_am[wc1x, wc1y, wc1z]
-    if ti.abs(current_val) < min_amplitude:
-        phase_sign = 1.0 if current_val >= 0.0 else -1.0
-        wave_field.psiL_am[wc1x, wc1y, wc1z] = phase_sign * min_amplitude
-
-
-@ti.func
-def interact_wc_drain(wave_field: ti.template()):  # type: ignore
-    """
-    TEST: Wave center as DRAIN - amplitude reduction
-    Absorbs wave energy at the wave center (WC) sphere surface
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-    """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-    wc_radius = 8  # radius in voxels
-    wc_radius_sq = wc_radius**2  # radius² = 8² voxels (≈ λ/4 for λ=30dx)
-
-    for i, j, k in ti.ndrange(
-        (wc1x - wc_radius - 1, wc1x + wc_radius + 2),
-        (wc1y - wc_radius - 1, wc1y + wc_radius + 2),
-        (wc1z - wc_radius - 1, wc1z + wc_radius + 2),
-    ):
-        dist_sq = (i - wc1x) ** 2 + (j - wc1y) ** 2 + (k - wc1z) ** 2
-        # Only process voxels on inner surface of sphere (r = radius-1)
-        if dist_sq <= wc_radius_sq:
-            wave_field.psiL_am[i, j, k] = 0.5 * wave_field.psiL_am[i, j, k]
-
-
-@ti.func
-def interact_wc_newmann(wave_field: ti.template()):  # type: ignore
-    """
-    TEST: Neumann boundary condition on wave center sphere surface
-    ∂ψ/∂n = 0 → copy outer values to inner surface (zero gradient)
-    This reflects waves WITHOUT phase inversion (soft/free-end reflection)
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-    """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-    wc_radius = 1  # radius in voxels
-    wc_radius_sq = wc_radius**2  # radius² = 8² voxels (≈ λ/4 for λ=30dx)
-    wc_radius_inner_sq = (wc_radius - 1) ** 2
-
-    for i, j, k in ti.ndrange(
-        (wc1x - wc_radius - 1, wc1x + wc_radius + 2),
-        (wc1y - wc_radius - 1, wc1y + wc_radius + 2),
-        (wc1z - wc_radius - 1, wc1z + wc_radius + 2),
-    ):
-        dist_sq = (i - wc1x) ** 2 + (j - wc1y) ** 2 + (k - wc1z) ** 2
-        # Only process voxels on inner surface of sphere (r = radius-1)
-        if wc_radius_inner_sq <= dist_sq <= wc_radius_sq:
-            # Find direction toward center and copy from outer neighbor
-            di = ti.cast(ti.math.sign(wc1x - i), ti.i32)
-            dj = ti.cast(ti.math.sign(wc1y - j), ti.i32)
-            dk = ti.cast(ti.math.sign(wc1z - k), ti.i32)
-            # Copy from the voxel just outside the sphere (Neumann: ∂ψ/∂n = 0)
-            outer_val = wave_field.psiL_am[i - di, j - dj, k - dk]
-            wave_field.psiL_am[i, j, k] = outer_val
-
-
-@ti.func
-def interact_wc_dirichlet(wave_field: ti.template()):  # type: ignore
-    """
-    TEST: Dirichlet boundary condition on wave center sphere surface
-    ψ = 0 →
-    This reflects waves WITH phase inversion (hard/fixed-end reflection)
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-    """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-    wc_radius = 8  # radius in voxels
-    wc_radius_sq = wc_radius**2  # radius² = 8² voxels (≈ λ/4 for λ=30dx)
-
-    for i, j, k in ti.ndrange(
-        (wc1x - wc_radius - 1, wc1x + wc_radius + 2),
-        (wc1y - wc_radius - 1, wc1y + wc_radius + 2),
-        (wc1z - wc_radius - 1, wc1z + wc_radius + 2),
-    ):
-        dist_sq = (i - wc1x) ** 2 + (j - wc1y) ** 2 + (k - wc1z) ** 2
-        # Only process voxels on inner surface of sphere (r = radius-1)
-        if dist_sq <= wc_radius_sq:
-            wave_field.psiL_old_am[i, j, k] = 0.0
-            wave_field.psiL_am[i, j, k] = 0.0
-            wave_field.psiL_new_am[i, j, k] = 0.0
-
-
-@ti.func
-def interact_wc_signal(wave_field: ti.template()):  # type: ignore
-    """
-    TEST: Wave center as SIGNAL INVERTER - phase inversion at WC sphere surface
-    Inverts wave phase at the wave center (WC) sphere surface
-
-    Args:
-        wave_field: WaveField instance containing displacement arrays and grid info
-    """
-    wc1x, wc1y, wc1z = wave_field.nx * 4 // 6, wave_field.ny * 4 // 6, wave_field.nz // 2
-    wc_radius = 0  # radius in voxels
-    wc_radius_sq = wc_radius**2  # radius² = 8² voxels (≈ λ/4 for λ=30dx)
-
-    for i, j, k in ti.ndrange(
-        (wc1x - wc_radius - 1, wc1x + wc_radius + 2),
-        (wc1y - wc_radius - 1, wc1y + wc_radius + 2),
-        (wc1z - wc_radius - 1, wc1z + wc_radius + 2),
-    ):
-        dist_sq = (i - wc1x) ** 2 + (j - wc1y) ** 2 + (k - wc1z) ** 2
-        # Only process voxels on inner surface of sphere (r = radius-1)
-        if dist_sq <= wc_radius_sq:
-            wave_field.psiL_am[i, j, k] = -wave_field.psiL_am[i, j, k]
+        wave_field.psiL_am[i, j, k] *= damping_factor
 
 
 # ================================================================
@@ -1257,7 +913,7 @@ def sample_avg_trackers(
 
 
 # ================================================================
-# FLUX MESH COLOR UPDATING
+# FLUX MESH VALUES UPDATING
 # ================================================================
 
 
@@ -1303,25 +959,37 @@ def update_flux_mesh_values(
             wave_field.fluxmesh_xy_colors[i, j] = colormap.get_blueprint_color(
                 freq_value, 0.0, trackers.avg_freq_rHz[None] * 2
             )
-            warp = freq_value / trackers.avg_freq_rHz[None] / 100 + wave_field.flux_mesh_planes[2]
+            warp = freq_value / trackers.avg_freq_rHz[None] / 100 + wave_field.flux_mesh_planes[
+                2
+            ] * (wave_field.nz / wave_field.max_grid_size)
             wave_field.fluxmesh_xy_vertices[i, j][2] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[2]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[2] * (wave_field.nz / wave_field.max_grid_size)
             )
         elif color_palette == 4:  # viridis
             wave_field.fluxmesh_xy_colors[i, j] = colormap.get_viridis_color(
                 ampL_value, 0, trackers.rms_ampL_am[None] * 2
             )
-            warp = ampL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[2]
+            warp = ampL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[
+                2
+            ] * (wave_field.nz / wave_field.max_grid_size)
             wave_field.fluxmesh_xy_vertices[i, j][2] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[2]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[2] * (wave_field.nz / wave_field.max_grid_size)
             )
         elif color_palette == 3:  # ironbow
             wave_field.fluxmesh_xy_colors[i, j] = colormap.get_ironbow_color(
                 ampT_value, 0, trackers.rms_ampT_am[None] * 2
             )
-            warp = ampT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[2]
+            warp = ampT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[
+                2
+            ] * (wave_field.nz / wave_field.max_grid_size)
             wave_field.fluxmesh_xy_vertices[i, j][2] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[2]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[2] * (wave_field.nz / wave_field.max_grid_size)
             )
         elif color_palette == 2:  # yellowgreen
             wave_field.fluxmesh_xy_colors[i, j] = colormap.get_yellowgreen_color(
@@ -1329,9 +997,13 @@ def update_flux_mesh_values(
                 -trackers.rms_ampL_am[None] * 2,
                 trackers.rms_ampL_am[None] * 2,
             )
-            warp = psiL_value / trackers.rms_ampL_am[None] / 1000 + wave_field.flux_mesh_planes[2]
+            warp = psiL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[
+                2
+            ] * (wave_field.nz / wave_field.max_grid_size)
             wave_field.fluxmesh_xy_vertices[i, j][2] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[2]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[2] * (wave_field.nz / wave_field.max_grid_size)
             )
         else:  # default to redblue (palette 1)
             wave_field.fluxmesh_xy_colors[i, j] = colormap.get_redblue_color(
@@ -1339,9 +1011,13 @@ def update_flux_mesh_values(
                 -trackers.rms_ampT_am[None] * 2,
                 trackers.rms_ampT_am[None] * 2,
             )
-            warp = psiT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[2]
+            warp = psiT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[
+                2
+            ] * (wave_field.nz / wave_field.max_grid_size)
             wave_field.fluxmesh_xy_vertices[i, j][2] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[2]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[2] * (wave_field.nz / wave_field.max_grid_size)
             )
 
     # ================================================================
@@ -1361,25 +1037,37 @@ def update_flux_mesh_values(
             wave_field.fluxmesh_xz_colors[i, k] = colormap.get_blueprint_color(
                 freq_value, 0.0, trackers.avg_freq_rHz[None] * 2
             )
-            warp = freq_value / trackers.avg_freq_rHz[None] / 100 + wave_field.flux_mesh_planes[1]
+            warp = freq_value / trackers.avg_freq_rHz[None] / 100 + wave_field.flux_mesh_planes[
+                1
+            ] * (wave_field.ny / wave_field.max_grid_size)
             wave_field.fluxmesh_xz_vertices[i, k][1] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[1]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[1] * (wave_field.ny / wave_field.max_grid_size)
             )
         elif color_palette == 4:  # viridis
             wave_field.fluxmesh_xz_colors[i, k] = colormap.get_viridis_color(
                 ampL_value, 0, trackers.rms_ampL_am[None] * 2
             )
-            warp = ampL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[1]
+            warp = ampL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[
+                1
+            ] * (wave_field.ny / wave_field.max_grid_size)
             wave_field.fluxmesh_xz_vertices[i, k][1] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[1]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[1] * (wave_field.ny / wave_field.max_grid_size)
             )
         elif color_palette == 3:  # ironbow
             wave_field.fluxmesh_xz_colors[i, k] = colormap.get_ironbow_color(
                 ampT_value, 0, trackers.rms_ampT_am[None] * 2
             )
-            warp = ampT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[1]
+            warp = ampT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[
+                1
+            ] * (wave_field.ny / wave_field.max_grid_size)
             wave_field.fluxmesh_xz_vertices[i, k][1] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[1]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[1] * (wave_field.ny / wave_field.max_grid_size)
             )
         elif color_palette == 2:  # yellowgreen
             wave_field.fluxmesh_xz_colors[i, k] = colormap.get_yellowgreen_color(
@@ -1387,9 +1075,13 @@ def update_flux_mesh_values(
                 -trackers.rms_ampL_am[None] * 2,
                 trackers.rms_ampL_am[None] * 2,
             )
-            warp = psiL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[1]
+            warp = psiL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[
+                1
+            ] * (wave_field.ny / wave_field.max_grid_size)
             wave_field.fluxmesh_xz_vertices[i, k][1] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[1]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[1] * (wave_field.ny / wave_field.max_grid_size)
             )
         else:  # default to redblue (palette 1)
             wave_field.fluxmesh_xz_colors[i, k] = colormap.get_redblue_color(
@@ -1397,9 +1089,13 @@ def update_flux_mesh_values(
                 -trackers.rms_ampT_am[None] * 2,
                 trackers.rms_ampT_am[None] * 2,
             )
-            warp = psiT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[1]
+            warp = psiT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[
+                1
+            ] * (wave_field.ny / wave_field.max_grid_size)
             wave_field.fluxmesh_xz_vertices[i, k][1] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[1]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[1] * (wave_field.ny / wave_field.max_grid_size)
             )
 
     # ================================================================
@@ -1419,25 +1115,37 @@ def update_flux_mesh_values(
             wave_field.fluxmesh_yz_colors[j, k] = colormap.get_blueprint_color(
                 freq_value, 0.0, trackers.avg_freq_rHz[None] * 2
             )
-            warp = freq_value / trackers.avg_freq_rHz[None] / 100 + wave_field.flux_mesh_planes[0]
+            warp = freq_value / trackers.avg_freq_rHz[None] / 100 + wave_field.flux_mesh_planes[
+                0
+            ] * (wave_field.nx / wave_field.max_grid_size)
             wave_field.fluxmesh_yz_vertices[j, k][0] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[0]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[0] * (wave_field.nx / wave_field.max_grid_size)
             )
         elif color_palette == 4:  # viridis
             wave_field.fluxmesh_yz_colors[j, k] = colormap.get_viridis_color(
                 ampL_value, 0, trackers.rms_ampL_am[None] * 2
             )
-            warp = ampL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[0]
+            warp = ampL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[
+                0
+            ] * (wave_field.nx / wave_field.max_grid_size)
             wave_field.fluxmesh_yz_vertices[j, k][0] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[0]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[0] * (wave_field.nx / wave_field.max_grid_size)
             )
         elif color_palette == 3:  # ironbow
             wave_field.fluxmesh_yz_colors[j, k] = colormap.get_ironbow_color(
                 ampT_value, 0, trackers.rms_ampT_am[None] * 2
             )
-            warp = ampT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[0]
+            warp = ampT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[
+                0
+            ] * (wave_field.nx / wave_field.max_grid_size)
             wave_field.fluxmesh_yz_vertices[j, k][0] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[0]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[0] * (wave_field.nx / wave_field.max_grid_size)
             )
         elif color_palette == 2:  # yellowgreen
             wave_field.fluxmesh_yz_colors[j, k] = colormap.get_yellowgreen_color(
@@ -1445,9 +1153,13 @@ def update_flux_mesh_values(
                 -trackers.rms_ampL_am[None] * 2,
                 trackers.rms_ampL_am[None] * 2,
             )
-            warp = psiL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[0]
+            warp = psiL_value / trackers.rms_ampL_am[None] / 100 + wave_field.flux_mesh_planes[
+                0
+            ] * (wave_field.nx / wave_field.max_grid_size)
             wave_field.fluxmesh_yz_vertices[j, k][0] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[0]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[0] * (wave_field.nx / wave_field.max_grid_size)
             )
         else:  # default to redblue (palette 1)
             wave_field.fluxmesh_yz_colors[j, k] = colormap.get_redblue_color(
@@ -1455,7 +1167,11 @@ def update_flux_mesh_values(
                 -trackers.rms_ampT_am[None] * 2,
                 trackers.rms_ampT_am[None] * 2,
             )
-            warp = psiT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[0]
+            warp = psiT_value / trackers.rms_ampT_am[None] / 100 + wave_field.flux_mesh_planes[
+                0
+            ] * (wave_field.nx / wave_field.max_grid_size)
             wave_field.fluxmesh_yz_vertices[j, k][0] = (
-                warp if warp_mesh else wave_field.flux_mesh_planes[0]
+                warp
+                if warp_mesh
+                else wave_field.flux_mesh_planes[0] * (wave_field.nx / wave_field.max_grid_size)
             )
