+++ openwave/xperiments/D_granule_based/_research/images/wave_amplitude_vs_radius.png
