+++ openwave/xperiments/D_granule_based/_xparameters/standing_wave.py
