--- openwave/xperiments/F_vector_field/_xparameters/the_queen.py
@@ -1,45 +1,0 @@
-"""
-XPERIMENT PARAMETERS
-
-This XPERIMENT showcases:
--
-"""
-
-UNIVERSE_EDGE = 1e-15  # m, universe edge length in meters
-
-XPARAMETERS = {
-    "meta": {
-        "X_NAME": "The Queen",
-        "DESCRIPTION": "Energy Wave Charging, Propagation and Interaction",
-    },
-    "camera": {
-        "INITIAL_POSITION": [1.17, 1.59, 0.81],  # [x, y, z] in normalized coordinates
-    },
-    "universe": {
-        "SIZE": [UNIVERSE_EDGE, UNIVERSE_EDGE, UNIVERSE_EDGE],  # m, simulation domain [x, y, z]
-        "TARGET_VOXELS": 2.5e7,  # Simulation voxel count (impacts performance)
-    },
-    "charging": {
-        "STATIC_BOOST": 1.00,  # One-Time charger amplitude boost multiplier
-    },
-    "ui_defaults": {
-        "SHOW_AXIS": False,  # Toggle to show/hide axis lines
-        "TICK_SPACING": 0.25,  # Axis tick marks spacing for position reference
-        "SHOW_GRID": False,  # Toggle to show/hide the voxel data-grid
-        "SHOW_EDGES": True,  # Toggle to show/hide universe edges
-        "FLUX_MESH_PLANES": [0.5, 0.5, 0.5],  # [x, y, z] positions relative to universe size
-        "SHOW_FLUX_MESH": 3,  # Flux Mesh toggle, 0: none, 1: xy, 2: xy+xz, 3: xy+xz+yz
-        "WARP_MESH": False,  # Toggle to enable/disable warp mesh effect
-        "SIM_SPEED": 1.0,  # Simulation speed multiplier
-        "PAUSED": False,  # Pause/Start simulation toggle
-    },
-    "color_defaults": {
-        "COLOR_THEME": "OCEAN",  # Choose color theme for rendering (OCEAN, DESERT, FOREST)
-        "COLOR_PALETTE": 2,  # redblue (1), yellowgreen (2), ironbow (3), viridis (4), blueprint (5)
-    },
-    "analytics": {
-        "INSTRUMENTATION": False,  # Toggle data acquisition and analytics
-        "EXPORT_VIDEO": False,  # Toggle frame image export to video directory
-        "VIDEO_FRAMES": 24,  # Target frame number to stop recording and finalize video export
-    },
-}
